+++ MovieQuiz/Services/ImageLoader.swift
@@ -7,11 +7,13 @@
 
 import Foundation
 
-protocol ImageLoading {
+protocol ImageLoaderProtocol {
+    
     func loadImageData(for movie: MostPopularMovie) throws -> Data
+    
 }
 
-struct ImageLoader: ImageLoading {
+struct ImageLoader: ImageLoaderProtocol {
     
     private let networkClient = NetworkClient()
