+++ MovieQuiz/Services/NetworkClient.swift
@@ -8,7 +8,9 @@
 import Foundation
 
 protocol NetworkRouting {
+    
     func fetch(url: URL, handler: @escaping (Result<Data, Error>) -> Void)
+    
 }
 
 struct NetworkClient: NetworkRouting {
